# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Our site
  version: 0.1.0
paths:
...
  /api/users/me/orders:
    get:
      tags:
      - Orders
      summary: Получить список заказов текущего пользователя
      operationId: list_my_orders_api_users_me_orders_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
          description: Фильтр по статусу заказа
          title: Status
        description: Фильтр по статусу заказа
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Смещение (offset)
          default: 0
          title: Offset
        description: Смещение (offset)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Лимит на страницу
          default: 50
          title: Limit
        description: Лимит на страницу
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    post:
      tags:
      - Orders
      summary: Создать новый заказ для текущего пользователя
      operationId: create_order_me_api_users_me_orders_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/orders:
    get:
      tags:
      - Orders
      summary: Получить список заказов пользователя по ID (Admin)
      operationId: list_orders_by_user_id_api_users__user_id__orders_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
          description: Фильтр по статусу заказа
          title: Status
        description: Фильтр по статусу заказа
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Смещение (offset)
          default: 0
          title: Offset
        description: Смещение (offset)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Лимит на страницу
          default: 50
          title: Limit
        description: Лимит на страницу
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    post:
      tags:
      - Orders
      summary: Создать новый заказ для пользователя по ID (Admin)
      operationId: create_order_by_user_id_api_users__user_id__orders_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/orders/{order_id}:
    get:
      tags:
      - Orders
      summary: Получить информацию о заказе текущего пользователя по ID
      operationId: get_order_me_api_users_me_orders__order_id__get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/me/orders/{order_id}/cancel:
    post:
      tags:
      - Orders
      summary: Отменить заказ текущего пользователя (пока он не одобрен)
      operationId: cancel_order_me_post_api_users_me_orders__order_id__cancel_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/orders/{order_id}:
    get:
      tags:
      - Orders
      summary: Получить информацию о заказе пользователя по ID (Admin)
      operationId: get_order_by_user_id_api_users__user_id__orders__order_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/orders:
    get:
      tags:
      - Orders
      summary: Получить список заказов
      operationId: list_orders_api_orders_get
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Фильтр по ID пользователя
          title: User Id
        description: Фильтр по ID пользователя
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
          description: Фильтр по статусу заказа
          title: Status
        description: Фильтр по статусу заказа
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Смещение (offset)
          default: 0
          title: Offset
        description: Смещение (offset)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Лимит на страницу
          default: 50
          title: Limit
        description: Лимит на страницу
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/orders/{order_id}:
    get:
      tags:
      - Orders
      summary: Получить информацию о заказе по ID
      operationId: get_order_api_orders__order_id__get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/orders/{order_id}/cancel:
    post:
      tags:
      - Orders
      summary: Отменить заказ
      operationId: cancel_order_api_orders__order_id__cancel_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/orders/{order_id}/approve:
    post:
      tags:
      - Orders
      summary: Подтвердить заказ
      operationId: approve_order_api_orders__order_id__approve_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/orders/{order_id}/reject:
    post:
      tags:
      - Orders
      summary: Отклонить заказ
      operationId: reject_order_api_orders__order_id__reject_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
components:
  schemas:
...
    CreateOrderRequest:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
          description: ID заказываемого товара
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
          description: Количество товара (должно быть > 0)
          default: 1
      type: object
      required:
      - product_id
      title: CreateOrderRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrderResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - offset
      - limit
      title: OrderListResponse
    OrderResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        product_id:
          type: string
          format: uuid
          title: Product Id
        quantity:
          type: integer
          title: Quantity
        unit_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Unit Price
        total_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Amount
        status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - user_id
      - product_id
      - quantity
      - unit_price
      - total_amount
      - status
      - created_at
      title: OrderResponse
    OrderStatus:
      type: string
      enum:
      - CREATED
      - APPROVED
      - REJECTED
      - CANCELLED
      title: OrderStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorDetailResponse:
      title: ErrorDetailResponse
      type: object
      properties:
        detail:
          title: Detail
          type: string
          description: Описание ошибки
      required:
      - detail
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Access Token (в заголовке `Authorization: Bearer <token>`
        или cookie `access_token`)'